pyproject.toml
[tool.poetry]
name = "thehalomod"
version = "0.1.0"
description = "The web-app for calculating Halo Model quantities"
authors = ["Steven Murray <steven.g.murray@asu.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
matplotlib = "^3.2.2"
hmf = "^3.2"
sentry_sdk = "^0.15.1"
django-analytical = "^2.6.0"
django-crispy-forms = "^1.9.1"
django-tabination = "^0.4.0"
django-range-slider = "^0.2.4"
numba = "^0.50.1"
setuptools_scm = "^4.1.2"
setuptools = "^47.3.1"
dill = "^0.3.2"
halomod = "^1.4.6-alpha.58"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
django-debug-toolbar = "^2.2"
django-extensions = "^3.0.4"
werkzeug = "^1.0.1"
whitenoise = "^5.2.0"
django-bootstrap-modal-forms = "^2.0.0"
djangorestframework = "^3.12.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
